containers/llm/requirements.txt
llama-cpp-python>=0.2.0
redis>=5.0.0
numpy>=1.21.0
pyyaml>=6.0
python-dotenv>=1.0.0
transformers>=4.30.0
sentencepiece>=0.1.99
